添加redis连接池

+++ src/main/java/cn/piao888/network/domain/bo/UploadBO.java
@@ -33,7 +33,7 @@ public class UploadBO {
     /**
      * 总切片数
      */
-    private Integer totalChunks;
+    private Long totalChunks;
     /**
      * 文件大小
      */
